Add to file and MSSQL support to SSP Training
This is a major update to SSP-Training. In addition to several bugs being fixed, the groovy scripts and the intermediate scripts have an option to put all output to a file. This is useful for creating the pre-compiled run once scripts in the main folder. Also, with a different command argument it is now possible to create and add the training data for MSSQL 2008.

While this is nearly complete the to file option for mssql isn't working at this time, and occaisonally  adding or editing an ssp-platform user for the first time after loading the training data, can cause a error however, subsequent attempts will work.